TRAIN-73 Refactor and Clean Current Trainees and Completed Trainees Pages
