+++ Frontend/src/components/DataGrid/DataGridPagination.tsx
@@ -1,0 +1,14 @@
+import React, { FC } from 'react';
+import Pagination from "./Pagination";
+import {GridPagination} from "@mui/x-data-grid";
+
+interface DataGridPaginationProps {}
+
+const DataGridPagination: FC<DataGridPaginationProps> = (props) => {
+
+    return (
+        <GridPagination ActionsComponent={Pagination} {...props} />
+    );
+};
+
+export default DataGridPagination;
