+++ Frontend/src/pages/university/CompletedTrainees/CompletedTrainees.css
@@ -1,13 +1,5 @@
-/*.dataGrid {*/
-/*    height: 90%;*/
-/*    width: 100%;*/
-/*    margin: auto;*/
-/*    margin-top: 2%;*/
-/*}*/
-
 .MuiDataGrid-columnHeaders {
     background-color: #ecf1f1;
-
 }
 
 .print-icon {
