+++ Frontend/src/pages/university/CompletedTrainees/components/EvaluationFormDialog.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from 'react';
+
+interface EvaluationFormDialogProps {}
+
+const EvaluationFormDialog: FC<EvaluationFormDialogProps> = (props) => {
+  
+  return (
+    <div>EvaluationFormDialog</div>
+  );
+};
+
+export default EvaluationFormDialog;
