+++ Frontend/src/pages/university/CompletedTrainees/types.ts
@@ -1,0 +1,7 @@
+export interface Row {
+    studentId: string;
+    Student: {
+        name: string;
+    };
+    count: string;
+}
