+++ Frontend/src/pages/university/CurrentTrainees/CurrentTrainees.css
@@ -1,13 +1,8 @@
-/*.dataGrid{*/
-/*    !*height: 90%; *!*/
-/*    width: 100%;*/
-/*    !*margin: auto;*!*/
-/*    !*margin-top: 2%;*!*/
-/* }*/
- .MuiDataGrid-columnHeaders{
-     background-color: #ecf1f1;
- }
- .manage-icon {
+.MuiDataGrid-columnHeaders {
+    background-color: #ecf1f1;
+}
+
+.manage-icon {
     font-size: 1.5em;
-  }
+}
