+++ Frontend/src/pages/university/CurrentTrainees/types.ts
@@ -1,0 +1,14 @@
+export interface Row {
+    id: string;
+    studentId: string;
+    companyBranchId: string;
+    Student: {
+        name: string;
+    };
+    CompanyBranch: {
+        location: string;
+        Company: {
+            name: string;
+        };
+    };
+}
